meta-pydmx/recipes-pydmx/pydmx/files/requirements.txt
PyQT5
customtkinter
darkdetect
packaging
requests
urllib3
idna
certifi
chardet
pillow
six
